Download spatial data from OD Surveys databases

The "read_map" function requires as parameter city, year, geometry (zones or municipalities) - the default is zone, and whether you want the harmonized database (variable names and reference/geodesic projection over the years, for the same city) or not - the default is the raw base.

read_map(city = "São Paulo", year = 2017, harmonize = FALSE, geometry = "zone")

Arguments

city

Character. City of reference. Defaults to "São Paulo".

year

Numeric. Year of reference in the format yyyy. Defaults to 1977.

harmonize

Logical. When FALSE (Default), the function returns the raw data. If TRUE, the function returns harmonized data to the same city, across all the years.

geometry

Character. The type of spatial data to be opened. Options include c("zone","district", "municipality").

Value

An "sf" "data.frame" object

Examples

if (FALSE) { # identical(tolower(Sys.getenv("NOT_CRAN")), "true")
library(odbr)

# return zone data from OD Surveys database as sf object at a given city and year
df <- read_map(
  city = "Sao Paulo",
  year = 2017,
  harmonize = FALSE,
  geometry = "zone"
)

#' # return district data from OD Surveys database as sf object at a given city and year
df <- read_map(
  city = "Sao Paulo",
  year = 2017,
  harmonize = FALSE,
  geometry = "district"
)

# return municipality data from OD Surveys database as sf object at a given city and year
df <- read_map(
  city = "Sao Paulo",
  year = 2017,
  harmonize = FALSE,
  geometry = "municipality"
)
}